# Conflicts:
#	common.py
#	dataclassform.py
#	models.py

+++ _admin/admin_faq.py
@@ -1,9 +1,9 @@
-from fastapi import APIRouter, Depends, File, Form, HTTPException, Request, UploadFile
-from fastapi.responses import RedirectResponse
+from fastapi import APIRouter, Depends, File, Form, Request, UploadFile
+from fastapi.responses import JSONResponse, RedirectResponse
 from fastapi.templating import Jinja2Templates
 from sqlalchemy.orm import Session
 from database import get_db
-import models
+from models import FaqMaster, Faq
 import shutil
 
 from common import *
@@ -12,13 +12,10 @@ from common import *
 router = APIRouter()
 templates = Jinja2Templates(directory=ADMIN_TEMPLATES_DIR)
 # 파이썬 함수 및 변수를 jinja2 에서 사용할 수 있도록 등록
-# TODO: base.html에서만 사용하는 함수이므로 이동이 필요함
 templates.env.globals["get_admin_menus"] = get_admin_menus
-# TODO: form에서 공용으로 사용하는 함수이므로 이동이 필요함
-templates.env.globals["generate_one_time_token"] = generate_one_time_token
-
-templates.env.globals["now"] = now
+templates.env.globals["generate_token"] = generate_token
 templates.env.globals["get_head_tail_img"] = get_head_tail_img
+templates.env.globals["now"] = now
 
 
 FAQ_MENU_KEY = "300700"
@@ -27,7 +24,7 @@ FAQ_MENU_KEY = "300700"
 @router.get("/faq_master_list")
 def faq_master_list(request: Request, db: Session = Depends(get_db)):
     """FAQ관리 목록"""
-    model = models.FaqMaster
+    model = FaqMaster
     request.session["menu_key"] = FAQ_MENU_KEY
 
     faq_masters = db.query(model).order_by(model.fm_order).all()
@@ -63,9 +60,10 @@ def faq_master_add(
     ):
     """FAQ관리 등록 처리"""
     # 토큰 검사
-    validate_token_or_raise(token)
+    if not compare_token(request, token, 'insert'):
+        raise AlertException(status_code=403, detail="토큰이 유효하지 않습니다.")
 
-    faq_master = models.FaqMaster(
+    faq_master = FaqMaster(
         fm_subject=fm_subject
         , fm_head_html=fm_head_html
         , fm_tail_html=fm_tail_html
@@ -98,7 +96,7 @@ def faq_master_update_form(fm_id: int, request: Request, db: Session = Depends(g
     """FAQ관리 수정 폼"""
     request.session["menu_key"] = FAQ_MENU_KEY
 
-    faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
+    faq_master = db.query(FaqMaster).filter(FaqMaster.fm_id == fm_id).first()
 
     return templates.TemplateResponse(
         "faq_master_form.html", {"request": request, "faq_master": faq_master}
@@ -124,9 +122,10 @@ def faq_master_update(
     ):
     """FAQ관리 수정 처리"""
     # 토큰 검사
-    validate_token_or_raise(token)
+    if not compare_token(request, token, 'update'):
+        raise AlertException(status_code=403, detail="토큰이 유효하지 않습니다.")
 
-    faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
+    faq_master = db.query(FaqMaster).filter(FaqMaster.fm_id == fm_id).first()
 
     faq_master.fm_subject = fm_subject
     faq_master.fm_head_html = fm_head_html
@@ -170,13 +169,14 @@ def faq_master_delete(
     # 토큰 검사
     # DELETE 요청일 경우, Body에 토큰이 없으므로 쿼리스트링에서 토큰을 얻는다.
     token = request.query_params.get("token")
-    validate_token_or_raise(token)
+    if not compare_token(request, token, 'delete'):
+        return JSONResponse(status_code=403, content={"message": "토큰이 유효하지 않습니다."})
     
-    faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
+    faq_master = db.query(FaqMaster).filter(FaqMaster.fm_id == fm_id).first()
     db.delete(faq_master)
     db.commit()
 
-    return {"message": "FAQ가 성공적으로 삭제되었습니다."}
+    return JSONResponse(status_code=200, content={"message": "FAQ가 성공적으로 삭제되었습니다."})
 
 
 @router.get("/faq_master_form/{fm_id}/list")
@@ -186,7 +186,7 @@ def faq_list(fm_id: int, request: Request, db: Session = Depends(get_db)):
     """
     request.session["menu_key"] = FAQ_MENU_KEY
 
-    faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
+    faq_master = db.query(FaqMaster).filter(FaqMaster.fm_id == fm_id).first()
     faqs = sorted(faq_master.faqs, key=lambda x: x.fa_order)
 
     return templates.TemplateResponse(
@@ -199,7 +199,7 @@ def faq_add_form(fm_id: int, request: Request, db: Session = Depends(get_db)):
     """FAQ관리 등록 폼"""
     request.session["menu_key"] = FAQ_MENU_KEY
 
-    faq_master = db.query(models.FaqMaster).filter(models.FaqMaster.fm_id == fm_id).first()
+    faq_master = db.query(FaqMaster).filter(FaqMaster.fm_id == fm_id).first()
 
     return templates.TemplateResponse(
         "faq_form.html", {"request": request, "faq_master": faq_master, "faq": None}
@@ -218,9 +218,10 @@ def faq_add(
     ):
     """FAQ관리 등록 처리"""
     # 토큰 검사
-    validate_token_or_raise(token)
+    if not compare_token(request, token, 'insert'):
+        raise AlertException(status_code=403, detail="토큰이 유효하지 않습니다.")
 
-    faq = models.Faq(
+    faq = Faq(
         fm_id=fm_id,
         fa_order=fa_order,
         fa_subject=fa_subject,
@@ -237,7 +238,7 @@ def faq_update_form(fa_id: int, request: Request, db: Session = Depends(get_db))
     """FAQ관리 등록 폼"""
     request.session["menu_key"] = FAQ_MENU_KEY
 
-    faq = db.query(models.Faq).filter(models.Faq.fa_id == fa_id).first()
+    faq = db.query(Faq).filter(Faq.fa_id == fa_id).first()
     faq_master = faq.faq_master
 
     return templates.TemplateResponse(
@@ -258,9 +259,10 @@ def faq_update(
     ):
     """FAQ관리 등록 처리"""
     # 토큰 검사
-    validate_token_or_raise(token)
+    if not compare_token(request, token, 'update'):
+        raise AlertException(status_code=403, detail="토큰이 유효하지 않습니다.")
 
-    faq = db.query(models.Faq).filter(models.Faq.fa_id == fa_id).first()
+    faq = db.query(Faq).filter(Faq.fa_id == fa_id).first()
 
     faq.fa_subject = fa_subject
     faq.fa_content = fa_content
@@ -280,10 +282,11 @@ async def faq_delete(
     # 토큰 검사
     # DELETE 요청일 경우, Body에 토큰이 없으므로 쿼리스트링에서 토큰을 얻는다.
     token = request.query_params.get("token")
-    validate_token_or_raise(token)
+    if not compare_token(request, token, 'delete'):
+        return JSONResponse(status_code=403, content={"message": "토큰이 유효하지 않습니다."})
 
-    faq = db.query(models.Faq).filter(models.Faq.fa_id == fa_id).first()
+    faq = db.query(Faq).filter(Faq.fa_id == fa_id).first()
     db.delete(faq)
     db.commit()
 
-    return {"message": "FAQ 항목이 성공적으로 삭제되었습니다."}
+    return JSONResponse(status_code=200, content={"message": "FAQ 항목이 성공적으로 삭제되었습니다."})
